Added Missing check for dependent namespaces
In the namespace YAML for the schema:

If a Namespace is dependent on another namespace, then its namespace schema will
include a namespace element in its schema list.  This will not include a `source`
which we touch to remove the file extension when embedding the cached schema.

This commit adds an extra check ensuring that `source` exists in the `schema`
element.