Changed heuristic for determining Class H5Type
Originally, we used the class hierarchy to determine which classes were groups
and which were datasets.  This required the schema to use the correct class
dependencies to determine which classes were what HDF5 type internally.

Apparently this will no longer work with the optional inclusion of commons.  This
fix adds more data to the Namespace class Map so that the file generator may
know what HDF5 type the class is supposed to be based on its parent group in
the schema itself.