Reverted to Released version of Nwb Schema 2.1.0
This branch was originally testing the update of the schema to 2.1.0 according
to a provided file.  As it turned out, the provided file was using an unstable
version of 2.1.0 whose version string was not updated (referred to as 3939e35,
its commit hash).

Due to the naming discrepancy, we've decided to first support the original
release of 2.1.0 while keeping the WIP changes from 3939e35.  Once the naming
issue is resolved, we will update to the new released schema version.