Added H5 Stub Superclasses
The original method for deriving H5 type of a class was not maintainable, so we
decided to use the schema instead.  However, the exporting class also needs this
information as well.

Two empty stub groups have been created as basic tags to identify whether or not
a given class is a Group or a Dataset: types.untyped.IsGroup and
types.untyped.IsDataset.