Minor update to README and add File Exchange badge
Altered the wording of the second Caveat section in the README in lieu of
instructions preferring pulling from nwb-schema.  Synchronizing the schema
with pynwb is still doable but is only really relevant if the user wishes
to use both pynwb and matnwb at once to create a file.

Added the new MatNWB file exchange badge.