Added Deflate Levels + Initial Writes to DataPipe
DataPipe now support deflate leves before binding.  The normal range is [0,9].
DataPipe considers the special value -1 as disabling compression. Note that
the compression isn't set until after export.

DataPipe now also supports Initial data writes on export in the event that the
user has no intention of appending to the dataset.  To do this, set the data
you wish to write in the `data` property of the DataPipe before Exporting.
Note that exporting disables `data` property and the data cannot be rewritten
without using a speciall overwriting `append` call.