Changed Namespace name filtering
Namespace on creation now filters the name to remove dashes.

Dashes on module names are not supported in Matlab so these must be stripped
before use.  Originally this was on a case-by-case basis but this change should
force the stripped namespace name to the correct format when querying from
Namespace.