Update Tests and minor schema fix
Fixed an issue where the test call to generateCore did not include hdmf_commons.
The call now includes hdmf_commons as a dependency.

Fixed a bunch of tests that relied on either VectorData, VectorIndex, or
ElementIdentifiers to be in the core namespace.  The references have been
corrected to the hdmf_common module.

The new schema also introduced the `required` keyword to datasets.  I've taken
this to imply overriding the `quantity` keyword.