Removed datenum Conversion for isodatetime
Originally, we simply used the DATENUM8601 function to produce date numbers to
then convert to datetime objects.  However, it is noted in the documentation for
the library that the date number format is not necessarily compatible with either
datestr or datetime.

We've instead opted to map the "token format provided by the function to the
MATLAB datetime input format.