Updated Extra Properties error
Unexpected extra property settings from a file usually indicate a schema
incompatibility between MatNWB and the file itself.  This usually occurs when
the file was not created with an embedded schema with which MatNWB could
conform itself to.  The error message has been updated to be more helpful in
this matter.

A utility function has also been added to aide in basic querying of the file
structure itself: `util.getSchemaVersion(filename)`
This function returns the nwb_version string in the root of the file.  The
assumption being that the attribute should remain in that spot of the file.