Refactoring DataPipe Start
DataPipe constructor split into three functions:

regular constructor(maxSize, axis) only takes the allocated size and an axis
to identify the append axis.

data queue fromData(data [, axis]) takes in data and creates a DataPipe from
it with the optional axis designation.  If no axis is provided, the assumed
axis is the first dimension (in MATLAB order).

fromFile(filename, path) returns a DataPipe already bound to the pre-created
dataset.  The axis must be set afterwards as the append axis is assumed to be
1.

Also added:
io.getMatType which does the inverse of io.getBaseType as it takes in a HDF5
tid and returns a MATLAB type name.

io.getMatTypeSize which return the size of the type given a MATLAB type name.

Draft types.untyped.datapipe.guessChunkSize which, in the future, will
implement the h5py method for estimating the best chunk size.

Modified:
io.parseDataset now updated to use the new fromFile call.