DataPipe constructor now warns if datatype invalid
the optional `datatype` keyword argument is used to specify the expected type
of the data when the data may not exist in memory.

When used with `data`, DataPipe prefers to use the `data` parameter to avoid
unnecessary type casting. This behavior is retained but a warning is added to
prevent ambiguity in types.