Fix metaclass failing to write compound references
Normally when a reference exists (any ObjectView or RegionView), it is added to
the `refs` list and immediately returned.

write_base does handle its own reference errors but when it returns, it actually
returns back into the export body, which causes errors since an attribute cannot be
written if the dataset isn't either.

This case only exists for compound datasets because we do a preliminary check earlier
in `export` for properties that are RegionViews/ObjectViews.

This should resolve the issue if the compound dataset contains a reference that simply
has not been written yet.