DataStub Fix data format issues
Sometimes the H5S groups blocks together. Originally it was assumed that these blocks
could only be a particular size in the first dimension, allowing for downstream
correction.

This has been changed so that cells are no longer returned by load_h5_style.
This is guaranteed to be a breaking change on other scripts but makes validation and
dimension correct much easier for load_mat_style downstream.