DataStub remove excess dim/0-index conversion
It is only necessary for DataStub to actually use C-style indexing and 0-index behavior
when interfacing with the low-level H5S functions. This conversion behavior is now
grouped with the low-level call itself and not at some arbitrary point beforehand.

This is to prevent data ambiguities when dealing with multi-dimension debugging.