Fix Compound tables with refs missing objects
There was a missing case where if the object was a dataset, its data a compound
datatype, and if the compound data has a reference object in it (one of ObjectView or
RegionView), then export() will fail because the metadata writers will be writing to
a non-existant dataset.

This commit checks if the returned refs from write_base are different, indicating that
nothing has actually been written, but deferred to the reference writing stage.