Fix addRow edge case with strings
When it comes to type validation, character matrices tend to become a pain. In this case
I figured it would be easier to allow scalar character arrays when adding to what may be
a cell array of strings. This was not allowed since the expected type was supposed to be
a "cell" instead of a "char".

A special case has been added that allows for row char arrays to be added to both char
matrices (as is default for numerical matrices) and cell arrays of characters.