Add support for other types in namespace cache
Originally only supported strings but since schemes.exportJson() actually requires
reading the cached schema directly, it was deemed simpler to actually just use the
numeric/logical information directly. This simplified a lot of the code that would
otherwise treat all data as character arrays.

Unfortunately, due to limitations with the YAML parser, rudimentary mapping rules were
added in the schema2matlab function to do basic type mapping. This is error-prone as
there is no way to verify the type of the original (i.e. "null" is read the exact same
way as [null] and "true" will always be converted to true).

Shape formatting was also changed to better accommodate actual numerical values in its
shape specifier.