Allow BlueprintPipe to report data size.
Originally, this was only possible for BoundPipes as they forwarded
DataStub data, but this method is also used when setting certain
data types (i.e ElectricalSeries). Since BlueprintPipe technically
can hold data, we do want to at least report what data it actually
has.

This is technically a hole as classes like ElectricalSeries don't
have continuous validation whenever the BlueprintPipe configuration
is set. This means that it is relatively simple to circumvent the
internal class's data shape this way.