Add parseConstrained ux warning
In the case where the user would wish to directly replace
a constrained property name with their own (usually scalar) value,
parseConstrained does not handle this properly, and assumes that
it will be handled in checkUnset. This is not true and should
throw a warning since the data is essentially thrown away.