Tests use correct file types.
Since we changed the default generation file location to the installation directory,
we want to maintain proper tests by forcing many of them to either not generate in the
home directory or only generate locally (in their respective working directories).

This is so that the installation directory is not polluted if `nwbtest` is called.