Specifically handle logical type concatenation
When concatenating to an empty double array "[]", MATLAB usually "adopts" the type from the new element
instead of attempting to cast to the double type.

As it turns out, logical values are a rare type where concatenating a "false" or "true" to an empty
array actually returns the double value of the empty array type. This added check fixes this
specific case.