Fix special duplicate H5D case
If the object's data is not a reference but has region/object references in its attributes, then
this error may occur because all reference corrections require calling `export` again.

Since we want to limit the possibility of object reference cycles, it's better to write any dataset
given that it's not reference data and just check for duplicates before doing so.