Improve Data Pipe Coverage (#557)
* add tests checking warnings

* expand init tests to include certain getters/setters

* use better testcase functions

* add bound pipe test

* Lower constructor errors to warnings

- Constructors should not normally error, but warnings should be emitted for the user to know
if certain features are unavailable. This is true for BoundPipe in case the pipe itself is not
chunked or a viable axis cannot be identified due to the strange shape of the datset chunks.
- simplified certain warning ids which were too noisy when it comes to identifying warnings.

* Elevate higher rank data to error when appending

If the data the user wishes to append has a higher rank than what is on the pipe, data loss
is required for appending. This is now an error because you probably never want to do that.

* add bound pipe non-chunked test

* shorten error ids regarding pipe properties